Release 1.8.0 (#87)
* Bugfix: Add string comparison for InstrumentType (#85)

* Bugfix: Add string comparison for InstrumentType

In older versions of this s4 packages, example, v1.6.0, the Instrument.instrument_type is a string.
So older s4 libraries does this sort of comparison:

bravo.instrument_type == "bravo"

But in the newer s4 packages, this comparison fails, because it's a string to object comparison. The user could update scripts, but something that is not feasible.

Simple solution: set the comparison, when it is a string, to compare against its name, which is more intuitive. 

This allows for backwards compatibility.
If doing an object-to-object comparison, it still does a comparison against URI, which is a source of truth.

* Update equality methods to use Union type hint instead of pipe character which was introduced in python v3.10

* Changed InstrumentType comparison overloading typehints to docstrings to match python2

* Add locked and archived properties to Researcher (#86)

Introduces 'locked' and 'archived' properties to the Researcher class for managing account lock status, including getter and setter methods. The 'archived' property serves as an alias for 'locked' to match UI terminology.

* Release 1.8.0

---------

Co-authored-by: Alan Lin <alin@billiontoone.com>
Co-authored-by: Brian Griffiths <bgriffiths@billiontoone.com>